Logger<T> and RootLogger moved to their own files.
